Move iRules-only verbs from tcl CLI to `f5 irule` sub-verbs
`event-order` and `event-info` are purely iRules concepts and now live
under the f5 CLI as `f5 irule event-order` and `f5 irule event-info`,
each defaulting to the f5-irules dialect.  Generally useful verbs that
happen to be common with iRules work (`command-info`, `convert`, `help`,
`diff`, the core analysis verbs) stay on tcl and continue to accept
`--dialect f5-irules` for iRules input.

Also drops the `irule` binary alias / `_default_dialect_for_prog` logic
from tcl_cli — there is no longer an `irule` symlink behaviour.  The
shell completion scripts, README, KCS docs, smoke-tests, and the zipapp
tcl entry point are all updated to match.
