+++ explorer/verbs/lookup.py
@@ -1,4 +1,8 @@
-"""Reference-lookup verbs: event-order, event-info, command-info, help."""
+"""Reference-lookup verbs: command-info, help.
+
+iRules-specific lookup verbs (``event-order``, ``event-info``) live on the
+``f5`` CLI under :mod:`explorer.verbs.f5.irule` instead.
+"""
 
 from __future__ import annotations
 
@@ -6,17 +10,12 @@ import argparse
 import json
 import sys
 
-from core.commands.registry.info import lookup_command_info, lookup_event_info
-from core.commands.registry.namespace_data import event_multiplicity, order_events_for_file
-from core.commands.registry.runtime import configure_signatures
+from core.commands.registry.info import lookup_command_info
 
 from ..pipeline import AVAILABLE_DIALECTS
 from ._registry import verb
 from ._utils import (
     TclCliError,
-    _add_input_arguments,
-    _combine_sources,
-    _read_input_documents,
     _write_text_output,
 )
 
@@ -99,49 +98,6 @@ _HELP_DIALECT_TERMS: dict[str, tuple[str, ...]] = {
         "tk",
     ),
 }
-
-
-@verb(
-    "event-order",
-    aliases=("eventorder",),
-    help="Show iRules events in canonical firing order.",
-)
-def _configure_event_order(
-    p: argparse.ArgumentParser, *, prog_name: str, default_dialect: str
-) -> None:
-    _add_input_arguments(p, include_output=True, default_dialect=default_dialect)
-    p.add_argument(
-        "--json",
-        action="store_true",
-        help="Emit event ordering as JSON.",
-    )
-    p.set_defaults(handler=_run_event_order)
-
-
-@verb(
-    "event-info",
-    aliases=("eventinfo",),
-    help="Look up iRules event metadata and valid commands.",
-)
-def _configure_event_info(
-    p: argparse.ArgumentParser, *, prog_name: str, default_dialect: str
-) -> None:
-    p.add_argument(
-        "event",
-        help="iRules event name (for example: HTTP_REQUEST).",
-    )
-    p.add_argument(
-        "--json",
-        action="store_true",
-        help="Emit event metadata as JSON.",
-    )
-    p.add_argument(
-        "--output",
-        "-o",
-        default="-",
-        help="Output path ('-' for stdout).",
-    )
-    p.set_defaults(handler=_run_event_info)
 
 
 @verb(
@@ -332,78 +288,6 @@ def _filter_catalogue_by_dialect(
 # ---------------------------------------------------------------------------
 # Handlers
 # ---------------------------------------------------------------------------
-
-
-def _run_event_order(args: argparse.Namespace) -> int:
-    documents = _read_input_documents(
-        args.inputs,
-        inline_sources=args.source,
-        package_paths=args.package_path,
-        recursive=not args.no_recursive,
-    )
-    configure_signatures(dialect=args.dialect)
-    source = _combine_sources(documents)
-
-    ordered = order_events_for_file(source)
-    events = [
-        {
-            "index": index,
-            "name": event_name,
-            "multiplicity": event_multiplicity(event_name),
-        }
-        for index, event_name in enumerate(ordered, start=1)
-    ]
-    payload = {
-        "count": len(events),
-        "dialect": args.dialect,
-        "events": events,
-    }
-    if args.json:
-        _write_text_output(args.output, json.dumps(payload, indent=2))
-        return 0
-
-    lines = [f"event order: {len(events)} event(s)"]
-    for item in events:
-        lines.append(f"  {item['index']}. {item['name']} ({item['multiplicity']})")
-    _write_text_output(args.output, "\n".join(lines))
-    return 0
-
-
-def _run_event_info(args: argparse.Namespace) -> int:
-    info = lookup_event_info(args.event, dialect="f5-irules")
-    payload = {
-        "event": info.event,
-        "known": info.known,
-        "deprecated": info.deprecated,
-        "multiplicity": info.multiplicity,
-        "description": info.description,
-        "side": info.side,
-        "transport": info.transport,
-        "impliedProfiles": list(info.implied_profiles),
-        "validCommandCount": info.valid_command_count,
-        "validCommands": list(info.valid_commands),
-    }
-
-    if args.json:
-        _write_text_output(args.output, json.dumps(payload, indent=2))
-        return 0 if info.known else 1
-
-    lines = [
-        f"event: {info.event}",
-        f"known: {'yes' if info.known else 'no'}",
-        f"deprecated: {'yes' if payload['deprecated'] else 'no'}",
-        f"multiplicity: {payload['multiplicity']}",
-    ]
-    if info.description:
-        lines.append(f"description: {info.description}")
-    lines.append(f"side: {payload['side']}")
-    if payload["transport"]:
-        lines.append(f"transport: {payload['transport']}")
-    if payload["impliedProfiles"]:
-        lines.append(f"profiles: {', '.join(str(item) for item in payload['impliedProfiles'])}")
-    lines.append(f"valid commands: {info.valid_command_count}")
-    _write_text_output(args.output, "\n".join(lines))
-    return 0 if info.known else 1
 
 
 def _run_command_info(args: argparse.Namespace) -> int:
